+++ packages/cli/__tests__/zosfiles/__integration__/compare/ds/__scripts__/bogus_flag.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+zowe zos-files compare ds "AT895452.TESTING.JCL(ECONTXTL)" "AT895452.TESTING.JCL(CONTXTLN)" --bogus-flag
+exit $?