+++ packages/cli/__tests__/zosfiles/__system__/compare/ds/Compare.ds.system.test.ts
@@ -10,7 +10,7 @@
 */
 
 import { Session } from "@zowe/imperative";
-import { Create, Delete, CreateDataSetTypeEnum } from "@zowe/cli";
+import { Create, Delete, CreateDataSetTypeEnum, Upload } from "@zowe/cli";
 import * as path from "path";
 import {ITestEnvironment, runCliScript} from "@zowe/cli-test-utils";
 import {TestEnvironment} from "../../../../../../../__tests__/__src__/environment/TestEnvironment";
@@ -22,12 +22,12 @@ let testEnvironment: ITestEnvironment<ITestPropertiesSchema>;
 let defaultSystem: ITestPropertiesSchema;
 let dsname: string;
 const TIMEOUT = 20000;
-
+const buffer = Buffer.from("buff");
 
 describe("Compare Data Sets", () => {
     beforeAll(async () => {
         testEnvironment = await TestEnvironment.setUp({
-            installPlugin: true,
+            // installPlugin: true,
             testName: "compare_data_sets",
             tempProfileTypes: ["zosmf"]
         });
@@ -52,6 +52,8 @@ describe("Compare Data Sets", () => {
     describe("Success scenarios", () => {
         beforeEach(async () => {
             await Create.dataSet(REAL_SESSION, CreateDataSetTypeEnum.DATA_SET_PARTITIONED, dsname);
+            await Upload.bufferToDataSet(REAL_SESSION, buffer, dsname+"(MEMB1)");
+            await Upload.bufferToDataSet(REAL_SESSION, buffer, dsname+"(MEMB2)");
         });
 
         afterEach(async () => {
@@ -59,17 +61,17 @@ describe("Compare Data Sets", () => {
         });
 
         it("should compare data sets", async () => {
-            const shellScript = path.join(__dirname, "__scripts__", "command", "command_compare_data_set.sh");
-            const response = runCliScript(shellScript, testEnvironment, [dsname]);
+            const shellScript = path.join(__dirname, "__scripts__/compare_data_set.sh");
+            const response = runCliScript(shellScript, testEnvironment, [dsname+"(MEMB1)", dsname+"(MEMB2)"]);
             expect(response.stderr.toString()).toBe("");
             expect(response.status).toBe(0);
             expect(response.stdout.toString()).toMatchSnapshot();
         }, TIMEOUT);
     });
     describe("Expected failures", () => {
-        it("should fail due to specified data set name does not existed", async () => {
-            const shellScript = path.join(__dirname, "__scripts__", "command", "command_compare_data_set.sh");
-            const response = runCliScript(shellScript, testEnvironment, [dsname + ".dummy"]);
+        it("should fail due to specified data set name does not exist", async () => {
+            const shellScript = path.join(__dirname, "__scripts__", "command", "command_compare_data_set.sh"); 
+            const response = runCliScript(shellScript, testEnvironment, [dsname+".dummy(MEMB1)", dsname+"(MEMB2)"]);
             expect(response.status).toBe(1);
             expect(response.stderr.toString()).toContain("Data set not found.");
         });