+++ packages/cli/__tests__/zosfiles/__unit__/compare/ds/Dataset.handler.unit.test.ts
@@ -11,7 +11,7 @@
 
 import { Get } from "@zowe/zos-files-for-zowe-sdk";
 import { UNIT_TEST_ZOSMF_PROF_OPTS } from "../../../../../../../__tests__/__src__/mocks/ZosmfProfileMock";
-import { DiffUtils } from "@zowe/imperative";
+import { DiffUtils, IDiffOptions } from "@zowe/imperative";
 
 describe("Compare data set handler", () => {
     describe("process method", () => {
@@ -27,11 +27,6 @@ describe("Compare data set handler", () => {
         let logMessage = "";
         let fakeSession: object;
 
-        // Mock the compare ds function
-        Get.dataSet = jest.fn(async (session) => {
-            fakeSession = session;
-            return Buffer.from("compared");
-        });
         // Mocked function references
         const profFunc = jest.fn((args) => {
             return {
@@ -43,7 +38,6 @@ describe("Compare data set handler", () => {
                 rejectUnauthorized: "fake",
             };
         });
-
         const processArguments = {
             arguments: {
                 $0: "fake",
@@ -81,6 +75,14 @@ describe("Compare data set handler", () => {
             }
         };
 
+        beforeEach(()=> {
+            // Mock the compare ds function
+            Get.dataSet = jest.fn(async (session) => {
+                fakeSession = session;
+                return Buffer.from("compared");
+            });
+            logMessage = "";
+        });
 
         it("should compare two data sets in terminal", async () => {
 
@@ -107,6 +109,45 @@ describe("Compare data set handler", () => {
             expect(apiMessage).toEqual("");
             expect(logMessage).toEqual("compared string");
             expect(DiffUtils.getDiffString).toHaveBeenCalledTimes(1);
+        });
+        it("should compare two data sets in terminal with --context-lines option", async () => {
+            let contextLinesArg: number = 2;
+            let processArgCopy: any = {
+                ...processArguments,
+                arguments:{
+                    ...processArguments.arguments,
+                    contextLines: contextLinesArg
+                }
+            };
+            let options: IDiffOptions = {
+                contextLinesArg, 
+                outputFormat: "terminal"
+            };
+
+            DiffUtils.getDiffString = jest.fn(async () => {
+                return "compared string";
+            });
+
+            try {
+                // Invoke the handler with a full set of mocked arguments and response functions
+                await handler.process(processArgCopy);
+            } catch (e) {
+                error = e;
+            }
+
+            expect(Get.dataSet).toHaveBeenCalledTimes(2);
+            expect(Get.dataSet).toHaveBeenCalledWith(fakeSession as any, dataSetName1, {
+                task: {
+                    percentComplete: 0,
+                    stageName: 0,
+                    statusMessage: "Retrieving second dataset"
+                }
+            });
+            expect(jsonObj).toMatchSnapshot();
+            expect(apiMessage).toEqual("");
+            expect(logMessage).toEqual("compared string");
+            expect(DiffUtils.getDiffString).toHaveBeenCalledTimes(1);
+            expect(DiffUtils.getDiffString).toHaveBeenCalledWith("compared", "compared", options)
         });
 
         it("should compare two data sets in browser", async () => {