+++ packages/cli/__tests__/zosfiles/__unit__/compare/lf-uss/LocalfileUss.handler.unit.test.ts
@@ -11,7 +11,8 @@
 
 import { Get } from "@zowe/zos-files-for-zowe-sdk";
 import { UNIT_TEST_ZOSMF_PROF_OPTS } from "../../../../../../../__tests__/__src__/mocks/ZosmfProfileMock";
-import { DiffUtils } from "@zowe/imperative";
+import { DiffUtils, IDiffOptions } from "@zowe/imperative";
+import { readFileSync } from "fs";
 
 describe("Compare local-file and uss-file handler", () => {
     describe("process method", () => {
@@ -28,12 +29,6 @@ describe("Compare local-file and uss-file handler", () => {
         let logMessage = "";
         let fakeSession: object;
 
-        // Mock the compare uss function
-        Get.USSFile = jest.fn(async (session) => {
-            fakeSession = session;
-            return Buffer.from("compared");
-        });
-
         // Mocked function references
         const profFunc = jest.fn((args) => {
             return {
@@ -45,7 +40,6 @@ describe("Compare local-file and uss-file handler", () => {
                 rejectUnauthorized: "fake",
             };
         });
-
         const processArguments = {
             arguments: {
                 $0: "fake",
@@ -83,7 +77,15 @@ describe("Compare local-file and uss-file handler", () => {
             }
         };
 
-
+        beforeEach(()=> {
+            // Mock the compare uss function
+            Get.USSFile = jest.fn(async (session) => {
+                fakeSession = session;
+                return Buffer.from("compared");
+            });
+            logMessage = "";
+        });
+        
         it("should compare local-file and uss-file in terminal", async () => {
 
             DiffUtils.getDiffString = jest.fn(async () => {
@@ -109,6 +111,46 @@ describe("Compare local-file and uss-file handler", () => {
             expect(apiMessage).toEqual("");
             expect(logMessage).toEqual("compared string");
             expect(DiffUtils.getDiffString).toHaveBeenCalledTimes(1);
+        });
+
+        it("should compare local-file and uss-file in terminal with --context-lines option", async () => {
+            let contextLinesArg: number = 2;
+            let processArgCopy: any = {
+                ...processArguments,
+                arguments:{
+                    ...processArguments.arguments,
+                    contextLines: contextLinesArg
+                }
+            };
+            let options: IDiffOptions = {
+                contextLinesArg, 
+                outputFormat: "terminal"
+            };
+
+            DiffUtils.getDiffString = jest.fn(async () => {
+                return "compared string";
+            });
+
+            try {
+                // Invoke the handler with a full set of mocked arguments and response functions
+                await handler.process(processArgCopy as any);
+            } catch (e) {
+                error = e;
+            }
+
+            expect(Get.USSFile).toHaveBeenCalledTimes(1);
+            expect(Get.USSFile).toHaveBeenCalledWith(fakeSession as any, ussFilePath, {
+                task: {
+                    percentComplete: 0,
+                    stageName: 0,
+                    statusMessage: "Retrieving uss file"
+                }
+            });
+            expect(jsonObj).toMatchSnapshot();
+            expect(apiMessage).toEqual("");
+            expect(logMessage).toEqual("compared string");
+            expect(DiffUtils.getDiffString).toHaveBeenCalledTimes(1);
+            expect(DiffUtils.getDiffString).toHaveBeenCalledWith(readFileSync(localFilePath).toString(), "compared", options);
         });
 
         it("should compare local-file and uss-file in browser", async () => {