+++ packages/cli/src/zosfiles/compare/lf-uss/LocalfileUss.handler.ts
@@ -76,7 +76,7 @@ export default class LocalfileUssHandler extends ZosFilesBaseHandler {
         let lfContentString = "";
         let ussContentString = "";
 
-        if (!commandParameters.arguments.seqnum) {
+        if (commandParameters.arguments.seqnum === false) {
             const seqnumlen = 8;
 
             const lfStringArray = lfContentBuf.toString().split("\n");